# openapi.yaml
openapi: 3.1.0
info:
  title: MedicalTime API
  version: 1.0.0
  description: API para lembretes de medicação, eventos de tomada, dispositivos, S.O.S. e autenticação.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    IntakeEventExpanded:
      type: object
      description: Evento de ingestão com dados do reminder embutidos para evitar round trips.
      properties:
        id: { type: string }
        medicationReminderId: { type: string }
        medicationScheduleId: { type: string, nullable: true }
        scheduledAt: { type: string, format: date-time }
        status: { type: string, enum: [PENDING, TAKEN, MISSED] }
        attempts: { type: integer }
        takenAt: { type: string, format: date-time, nullable: true }
        reminder:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            photoUrl: { type: string, nullable: true }
          required: [id, name]
        schedule:
          type: object
          nullable: true
          properties:
            id: { type: string }
            ingestionTimeMinutes: { type: integer }
          required: [id, ingestionTimeMinutes]
      required:
        [id, medicationReminderId, scheduledAt, status, attempts, reminder]
...
    EmergencyContact:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        priority: { type: integer }
        isActive: { type: boolean }
      required: [id, name, phone, priority, isActive]
    EmergencyContactCreate:
      type: object
      required: [name, phone]
      properties:
        name: { type: string }
        phone: { type: string }
        priority: { type: integer }
    EmergencyContactUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        priority: { type: integer }
        isActive: { type: boolean }
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }
          required: [code, message]
      required: [error]
  responses:
...
    NotFoundError:
      description: Não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            generico:
              summary: Recurso inexistente
              value:
                error:
                  code: NOT_FOUND
                  message: "Recurso não encontrado"
    RateLimitError:
      description: Rate limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            excesso:
              summary: Muitas requisições
              value:
                error:
                  code: RATE_LIMIT
                  message: "Muitas requisições. Tente novamente mais tarde."
                  details:
                    retryAfterSeconds: 42
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validacao:
              summary: Falha de validação de campos
              value:
                error:
                  code: BAD_REQUEST
                  message: "Falha de validação"
                  details:
                    issues:
                      - path: ["email"]
                        message: "E-mail inválido"
                      - path: ["password"]
                        message: "Mínimo de 6 caracteres"
...
    InternalServerError:
      description: Erro interno
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            generico:
              summary: Erro inesperado
              value:
                error:
                  code: INTERNAL_ERROR
                  message: "Erro interno inesperado"
security:
  - bearerAuth: []
paths:
...
  /intakes:
    get:
      operationId: listIntakeEvents
      summary: Listar eventos de tomada no intervalo (array simples)
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: hours
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, TAKEN, MISSED] }
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IntakeEventExpanded" }
              examples:
                proximas4h:
                  summary: Próximas 4 horas
                  value:
                    - id: intk_1
                      medicationReminderId: rem_1
                      medicationScheduleId: sch_1
                      scheduledAt: 2025-09-16T19:30:00.000Z
                      status: PENDING
                      attempts: 0
                      reminder:
                        id: rem_1
                        name: Vitamina D
                        photoUrl: null
                      schedule:
                        id: sch_1
                        ingestionTimeMinutes: 510
                    - id: intk_2
                      medicationReminderId: rem_2
                      medicationScheduleId: sch_5
                      scheduledAt: 2025-09-16T20:15:00.000Z
                      status: PENDING
                      attempts: 0
                      reminder:
                        id: rem_2
                        name: Ômega 3
                        photoUrl: null
                      schedule:
                        id: sch_5
                        ingestionTimeMinutes: 600
        "400": { $ref: "#/components/responses/BadRequestError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /intakes/history:
    get:
      operationId: listIntakeHistory
      summary: Histórico de eventos recentes
      parameters:
        - in: query
          name: days
          schema: { type: string }
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IntakeEventExpanded" }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /intakes/{id}/taken:
    post:
      operationId: markIntakeTaken
      summary: Marcar evento como tomado
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Ok }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /emergency/emergency-contacts:
    get:
      operationId: listEmergencyContacts
      summary: Listar contatos de emergência (array simples)
      responses:
        "200":
          description: Lista de contatos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EmergencyContact" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createEmergencyContact
      summary: Criar contato de emergência
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmergencyContactCreate" }
      responses:
        "201": { description: Criado }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /emergency/emergency-contacts/{id}:
    patch:
      operationId: updateEmergencyContact
      summary: Atualizar contato de emergência
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmergencyContactUpdate" }
      responses:
        "200": { description: Ok }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteEmergencyContact
      summary: Remover contato de emergência
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removido }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /emergency/sos:
    post:
      operationId: triggerSOS
      summary: Disparar SOS
      description: Dispara fluxo de notificação para contatos de emergência
      tags: [Emergency]
      x-codeSamples:
        - lang: curl
          label: Curl exemplo
          source: |
            curl -X POST \
              -H "Authorization: Bearer <TOKEN>" \
              http://localhost:3000/emergency/sos
      responses:
        "200": { description: Ok }
        "429": { $ref: "#/components/responses/RateLimitError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /dev/test-sos:
    post:
      operationId: devTestSOS
      summary: Disparar SOS (dev)
      x-codeSamples:
        - lang: curl
          label: Curl dev
          source: |
            curl -X POST http://localhost:3000/dev/test-sos
      responses:
        "200": { description: Ok }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /dev/test-alarm:
    post:
      operationId: devTestAlarm
      summary: Disparar alarme (dev)
      x-codeSamples:
        - lang: curl
          label: Curl dev alarm
          source: |
            curl -X POST http://localhost:3000/dev/test-alarm
      responses:
        "200": { description: Ok }
        "500": { $ref: "#/components/responses/InternalServerError" }